asciidoc3: self.newline replaced by '\n'...
No need to play with different newline sequences when using unicode
strings and the unified newlines. The change automagically fixed the
extra newlines generated to the HTML. The bad sequence (on Windows)
was '\r\r\n' (binary), now it is correctly '\r\n'. The reason for the
earlier output was not sniffed.