fix asciidoc.conf -- change the newline sequence to \n (was \r\n)
On Windows, the \r\n caused strange output line ends because of the
text mode of the output file.

The explicit newline sequence was probably important in Python 2.x
where files were open in binary mode. However, the asciidoc3 uses
UNICODE strings and the open() for the output should always use the
text mode. The usage of the explicitly used newline should probably
be removed completely